source/ucp-s3-excise-queue-processor/requirements.txt
awswrangler==3.5.1
pandas-stubs
coverage
testcontainers
botocore
boto3-stubs-lite[s3,glue,dynamodb]